Fix listContents() when includes a directory
What does this Pull Request Do?
-------------------------------
* Modified the `normalizeListingObject()` method. Identify the type of the object to normalize. If it's a 'file' the method will still return the same as before, if it's a 'dir' we take out `visibility` and `size` from the returned array.

How should this be manually tested?
-----------------------------------
A test was included with a similar output you can get in a server, which doesn't include the 'size'.

Any background context you want to provide?
-------------------------------------------
Currently there's a php `Notice` when trying to list the content of a directory with sub-directories.

What are the relevant tickets?
------------------------------
* [Undefined index: size when trying to get the listing of a directories with subdirectories · Issue #2 · thephpleague/flysystem-sftp](https://github.com/thephpleague/flysystem-sftp/issues/2)

Any extra info?
---------------
This is a minor bug. Please, give it a try and let me know.