added URL setter and getter
One might want to have access to a full public url for a file if it is available (i.e. sftp files to another server to be referenced publicly).  Also, when using this package with Laravel, most other adapters have a getUrl() method to allow the user to fetch a public url to the files if available.