Bugfix: Directories or files named '0' do not show up in listContents
When the filenames come from the phpseclib\Net\SFTP client they are passed
as array keys. If that filename is just '0' php handles it as an int.
In ContentListingFormatter there is a special case that handles !== '0',
since filename is an int in this case it does not match and is considered empty.

Just making sure filename is always a string solves this.